cleanup

+++ src/dll/zenwinx/string.c
@@ -619,14 +619,10 @@ int winx_bytes_to_hr(ULONGLONG bytes, int digits, char *buffer, int length)
     for(n = bytes, m = 1, i = 0; n >> 10; n >>= 10, m <<= 10, i++){}
     r = bytes - n * m;
     
-    /* Win DDK cannot convert ULONGLONG to double directly, */
-    /* but now it's safe to convert both r and m through LONGLONG */
-    rd = (double)(LONGLONG)r / (double)(LONGLONG)m;
+    rd = (double)r / (double)m;
     if(rd >= 1) rd = 0.999999999999999;
     rd *= pow(10, digits);
-    /* convertion to LONGLONG is needed for MinGW */
-    /* it is safe, but may lower a highest possible precision */
-    r = (ULONGLONG)(LONGLONG)rd;
+    r = (ULONGLONG)rd;
     
     if(digits == 0){
         result = _snprintf(buffer, length - 1, "%I64u %s", n, suffixes[i]);
